remove ComboButton component
