--- packages/react-components/react-combobox/src/ComboButton.ts
@@ -1,1 +1,0 @@
-export * from './components/ComboButton/index';
